Add publisher in Sticky_Leg_PLugin

+++ sar_env/SAR_Base_Interface.py
@@ -259,7 +259,7 @@ class SAR_Base_Interface(Node):
         
         return self.t
 
-    def sendCmd(self,action,cmd_vals=[0.0,0.0,0.0],cmd_flag=1):
+    def sendCmd(self,action,cmd_vals=[0.0,0.0,0.0],cmd_flag=1.0):
         """Sends commands to SAR_DC->Controller via rosservice call
 
         Args:
@@ -520,17 +520,17 @@ class SAR_Base_Interface(Node):
 
     def handle_Pos_Cmd(self):
         cmd_vals = self.userInput("Set desired position values (x,y,z): ",float)
-        cmd_flag = self.userInput("Pos control On/Off (1,0): ",int)
+        cmd_flag = self.userInput("Pos control On/Off (1,0): ",float)
         self.sendCmd("Pos",cmd_vals,cmd_flag)
 
     def handle_Vel_Cmd(self):
         cmd_vals = self.userInput("Set desired velocity values (x,y,z): ",float)
-        cmd_flag = self.userInput("Vel control On/Off (1,0): ",int)
+        cmd_flag = self.userInput("Vel control On/Off (1,0): ",float)
         self.sendCmd("Vel",cmd_vals,cmd_flag)
 
     def handle_Ang_Accel(self):
         cmd_vals = self.userInput("Set desired angular acceleration values (x,y,z): ",float)
-        cmd_flag = self.userInput("Ang_Accel control On/Off (1,0): ",int)
+        cmd_flag = self.userInput("Ang_Accel control On/Off (1,0): ",float)
         self.sendCmd("Ang_Accel",cmd_vals,cmd_flag)
 
     def handle_Policy(self):
@@ -549,9 +549,9 @@ class SAR_Base_Interface(Node):
         
     def handle_P2P_traj(self):
         x_d = self.userInput("Desired position (x,y,z):",float)
-        self.sendCmd('P2P_traj',cmd_vals=[np.nan,x_d[0],0.5],cmd_flag=0)
-        self.sendCmd('P2P_traj',cmd_vals=[np.nan,x_d[1],0.5],cmd_flag=1)
-        self.sendCmd('P2P_traj',cmd_vals=[np.nan,x_d[2],0.5],cmd_flag=2)
+        self.sendCmd('P2P_traj',cmd_vals=[np.nan,x_d[0],0.5],cmd_flag=0.0)
+        self.sendCmd('P2P_traj',cmd_vals=[np.nan,x_d[1],0.5],cmd_flag=1.0)
+        self.sendCmd('P2P_traj',cmd_vals=[np.nan,x_d[2],0.5],cmd_flag=2.0)
         self.sendCmd('Activate_traj',cmd_vals=[1.0,1.0,1.0])
 
     def handle_Global_Vel_traj(self):
@@ -566,8 +566,8 @@ class SAR_Base_Interface(Node):
         V_B_O = [Vx,Vy,Vz]
 
         ## EXECUTE TRAJECTORY
-        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[0],self.TrajAcc_Max[0],self.TrajJerk_Max[0]],cmd_flag=0)
-        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[2],self.TrajAcc_Max[2],self.TrajJerk_Max[2]],cmd_flag=2)
+        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[0],self.TrajAcc_Max[0],self.TrajJerk_Max[0]],cmd_flag=0.0)
+        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[2],self.TrajAcc_Max[2],self.TrajJerk_Max[2]],cmd_flag=2.0)
         self.sendCmd('Activate_traj',cmd_vals=[1.0,0.0,1.0])
 
     def handle_Rel_Vel_traj(self):
@@ -584,8 +584,8 @@ class SAR_Base_Interface(Node):
         V_B_O = self.R_PW(np.array([V_tx,V_ty,V_perp]),self.Plane_Angle_rad)
 
         ## EXECUTE TRAJECTORY
-        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[0],self.TrajAcc_Max[0],self.TrajJerk_Max[0]],cmd_flag=0)
-        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[2],self.TrajAcc_Max[2],self.TrajJerk_Max[2]],cmd_flag=2)
+        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[0],self.TrajAcc_Max[0],self.TrajJerk_Max[0]],cmd_flag=0.0)
+        self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[2],self.TrajAcc_Max[2],self.TrajJerk_Max[2]],cmd_flag=2.0)
         self.sendCmd('Activate_traj',cmd_vals=[1,0,1])
 
     def handle_Impact_traj(self):
@@ -614,9 +614,9 @@ class SAR_Base_Interface(Node):
         print(YELLOW,f"Start Position: ({r_B_O[0]:.2f},{self.r_B_O[1]:.2f},{r_B_O[2]:.2f})",RESET)
         str_input = self.userInput("Approve start position (y/n): ",str)
         if str_input == 'y':
-            self.sendCmd('P2P_traj',cmd_vals=[np.nan,r_B_O[0],0.5],cmd_flag=0)
-            self.sendCmd('P2P_traj',cmd_vals=[np.nan,r_B_O[1],0.5],cmd_flag=1)
-            self.sendCmd('P2P_traj',cmd_vals=[np.nan,r_B_O[2],0.5],cmd_flag=2)
+            self.sendCmd('P2P_traj',cmd_vals=[np.nan,r_B_O[0],0.5],cmd_flag=0.0)
+            self.sendCmd('P2P_traj',cmd_vals=[np.nan,r_B_O[1],0.5],cmd_flag=1.0)
+            self.sendCmd('P2P_traj',cmd_vals=[np.nan,r_B_O[2],0.5],cmd_flag=2.0)
             self.sendCmd('Activate_traj',cmd_vals=[1.0,1.0,1.0])
         else:
             raise Exception("Start position not approved")
@@ -624,13 +624,13 @@ class SAR_Base_Interface(Node):
         ## POLICY SENDING
         cmd_vals = self.userInput("Set desired (Tau,AngAcc) Policy: ",float)
         cmd_vals.append(-100.0) # Append extra value to match framework
-        self.sendCmd('Policy',cmd_vals,cmd_flag=0)
+        self.sendCmd('Policy',cmd_vals,cmd_flag=0.0)
 
         ## APPROVE FLIGHT
         str_input = self.userInput("Approve flight (y/n): ",str)
         if str_input == 'y':
-            self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[0],self.TrajAcc_Max[0],self.TrajJerk_Max[0]],cmd_flag=0)
-            self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[2],self.TrajAcc_Max[2],self.TrajJerk_Max[2]],cmd_flag=2)
+            self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[0],self.TrajAcc_Max[0],self.TrajJerk_Max[0]],cmd_flag=0.0)
+            self.sendCmd('Const_Vel_traj',cmd_vals=[V_B_O[2],self.TrajAcc_Max[2],self.TrajJerk_Max[2]],cmd_flag=2.0)
             self.sendCmd('Activate_traj',cmd_vals=[1.0,0.0,1.0])
         else:
             raise Exception("Flight not approved")
@@ -641,7 +641,7 @@ class SAR_Base_Interface(Node):
     def handle_Arm_Quad(self):
 
         #cmd_flag = self.userInput("Arm Quad On/Off (1,0): ",int)
-        cmd_flag = self.userInput("Arm Quad On/Off (1,0): ",int)
+        cmd_flag = self.userInput("Arm Quad On/Off (1,0): ",float)
         
         #For experiment
         #self.sendCmd("Load_Params")
@@ -653,7 +653,7 @@ class SAR_Base_Interface(Node):
     
     def handle_Tumble_Detect(self):
 
-        cmd_flag = self.userInput("Tumble Detection On/Off (1,0): ",int)
+        cmd_flag = self.userInput("Tumble Detection On/Off (1,0): ",float)
         self.sendCmd('Tumble_Detect',cmd_vals=[1.0,1.0,1.0],cmd_flag=cmd_flag)
 
     def handle_Load_Params(self):
